+++ lib/src/persistence/persistence_exception.dart
@@ -9,6 +9,13 @@ class PersistenceException implements Exception {
   PersistenceException.closed(DatabaseConnection connection)
       : this('Connection closed: $connection');
 
+  PersistenceException.internal(String message)
+      : this('A DataHub internal error occurred! This error should not have '
+            'happened and it is very likely, that this error is not your '
+            'fault.\nPlease report this error, this message and instructions '
+            'to reproduce it to the DataHub Issue Tracker.\n'
+            'Details: $message');
+
   @override
   String toString() => 'PersistenceException: $message';
 }