+++ lib/src/persistence/query/query_source.dart
@@ -7,7 +7,7 @@ import 'package:datahub/persistence.dart';
 abstract class QuerySource<T> {
   const QuerySource();
 
-  T map(List<QueryResult> results);
+  T? map(List<QueryResult> results);
 }
 
 //TODO use filter objects for join
@@ -23,6 +23,8 @@ class BeanJoin<TDao> {
 abstract class JoinedQuerySource {
   DataBean get main;
 
+  bool get innerJoin;
+
   List<BeanJoin> get joins;
 }
 
@@ -33,22 +35,37 @@ class TupleJoinQuerySource<Ta, Tb> extends QuerySource<Tuple<Ta, Tb>>
   final BeanJoin<Tb> joinB;
 
   @override
+  final innerJoin = !TypeCheck<Null>().isSubtypeOf<Tb>();
+
+  @override
   List<BeanJoin> get joins => [joinB];
 
   TupleJoinQuerySource(this.main, this.joinB);
 
   @override
   Tuple<Ta, Tb> map(List<QueryResult> results) {
-    return Tuple(main.map(results), joinB.bean.map(results));
+    final mainResult = main.map(results);
+    if (mainResult == null) {
+      throw PersistenceException.internal(
+          'MainResult of TupleJoinQuerySource is null.');
+    }
+
+    final resultB = joinB.bean.map(results);
+    if (innerJoin && resultB == null) {
+      throw PersistenceException.internal(
+          'ResultB of TupleJoinQuerySource is null but innerJoin == true.');
+    }
+
+    return Tuple(mainResult, resultB as Tb);
   }
 
   /// Create [JoinedQuerySource] with this and the following join.
   ///
   /// This enabled chaining of join() calls.
-  TripleJoinQuerySource<Ta, Tb, TDao> join<TDao>(
+  TripleJoinQuerySource<Ta, Tb, Tc> join<Tc, TDao extends Tc>(
       DataBean<TDao> other, DataField mainField, DataField otherField,
       {CompareType type = CompareType.equals}) {
-    return TripleJoinQuerySource(
+    return TripleJoinQuerySource<Ta, Tb, Tc>(
       main,
       joinB,
       BeanJoin<TDao>(other, mainField, type, otherField),
@@ -64,16 +81,38 @@ class TripleJoinQuerySource<Ta, Tb, Tc> extends QuerySource<Triple<Ta, Tb, Tc>>
   final BeanJoin<Tc> joinC;
 
   @override
+  final innerJoin = !TypeCheck<Null>().isSubtypeOf<Tb>() &&
+      !TypeCheck<Null>().isSubtypeOf<Tc>();
+
+  @override
   List<BeanJoin> get joins => [joinB, joinC];
 
   TripleJoinQuerySource(this.main, this.joinB, this.joinC);
 
   @override
   Triple<Ta, Tb, Tc> map(List<QueryResult> results) {
+    final mainResult = main.map(results);
+    if (mainResult == null) {
+      throw PersistenceException.internal(
+          'MainResult of TripleJoinQuerySource is null.');
+    }
+
+    final resultB = joinB.bean.map(results);
+    if (innerJoin && resultB == null) {
+      throw PersistenceException.internal(
+          'ResultB of TripleJoinQuerySource is null but innerJoin == true.');
+    }
+
+    final resultC = joinC.bean.map(results);
+    if (innerJoin && resultC == null) {
+      throw PersistenceException.internal(
+          'ResultC of TripleJoinQuerySource is null but innerJoin == true.');
+    }
+
     return Triple(
-      main.map(results),
-      joinB.bean.map(results),
-      joinC.bean.map(results),
+      mainResult,
+      resultB as Tb,
+      resultC as Tc,
     );
   }
 }